ОБъединить все в одно приложение
