+++ api_yamdb/api/urls.py
@@ -1,7 +1,6 @@
-from comments.views import CommentViewSet, ReviewModelViewSet
 from django.urls import include, path
 from rest_framework import routers
-from titles.views import CategoriesViewSet, GenresViewSet, TitlesViewSet
+from reviews.views import CategoriesViewSet, GenresViewSet, TitlesViewSet, CommentViewSet, ReviewModelViewSet
 
 v1_router = routers.DefaultRouter()
 v1_router.register(r'titles', TitlesViewSet)
@@ -17,7 +16,6 @@ v1_router.register(
     ReviewModelViewSet,
     basename="ReviewModel"
 )
-
 
 urlpatterns = [
     path('v1/', include(v1_router.urls)),
