--- api_yamdb/comments/models.py
@@ -1,59 +1,0 @@
-from django.contrib.auth import get_user_model
-from django.core.validators import MaxValueValidator, MinValueValidator
-from django.db import models
-from titles.models import TitlesModel
-
-User = get_user_model()
-
-
-class ReviewModel(models.Model):
-    title_id = models.ForeignKey(
-        TitlesModel,
-        verbose_name='Произведение',
-        on_delete=models.CASCADE,
-        related_name='reviews',
-    )
-    text = models.TextField('Текст отзыва')
-    author = models.ForeignKey(
-        User,
-        verbose_name='Автор отзыва',
-        on_delete=models.CASCADE,
-        related_name='reviews',
-        null=True,
-    )
-    score = models.IntegerField(
-        'Рейтинг', validators=[MinValueValidator(1), MaxValueValidator(10)]
-    )
-    pub_date = models.DateTimeField(
-        'Дата добавления', auto_now_add=True, db_index=True
-    )
-
-    class Meta:
-        verbose_name = 'Отзыв'
-        verbose_name_plural = 'Отзывы'
-
-    def __str__(self):
-        return self.title_id.name
-
-
-class CommentModel(models.Model):
-    author = models.ForeignKey(
-        User,
-        verbose_name='Автор комментария',
-        on_delete=models.CASCADE,
-        related_name='comments',
-    )
-    review_id = models.ForeignKey(
-        ReviewModel,
-        verbose_name='Отзыв',
-        on_delete=models.CASCADE,
-        related_name='comments',
-    )
-    text = models.TextField('Текст комментария')
-    pub_date = models.DateTimeField(
-        'Дата добавления', auto_now_add=True, db_index=True
-    )
-
-    class Meta:
-        verbose_name = 'Комментарий'
-        verbose_name_plural = 'Комментарии'
