+++ api_yamdb/reviews/apps.py
@@ -1,7 +1,6 @@
 from django.apps import AppConfig
 
 
-class CommentsConfig(AppConfig):
+class ReviewsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'comments'
-    verbose_name = 'Отзывы'
+    name = 'reviews'
