--- api_yamdb/titles/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class TitlesConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'titles'
-    verbose_name = 'Произведения'
